bugfix: fix classic item icons, fix classic favorites feat: new UI for mythic raid leaderboards left panel

+++ app/src/main/java/com/BlizzardArmory/model/common/Media.kt
@@ -8,5 +8,5 @@ import com.google.gson.annotations.SerializedName
 data class Media(
 
     @SerializedName("key") val key: Key,
-    @SerializedName("id") val id: Int
+    @SerializedName("id") val id: Long
 )
